% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movres.R
\name{approxNormalOrder}
\alias{approxNormalOrder}
\alias{approxNormalOrder2}
\title{Auxiliary for Preparing Discrete Distribution
used to approximating Standard Normal Distribution}
\usage{
approxNormalOrder(m)

approxNormalOrder2(m)
}
\arguments{
\item{m}{int, the number of order statistics used}
}
\value{
A numeric matrix with first column is support of discrete
distribution and second column is corresponding p.m.f..
}
\description{
Auxiliary for preparing discrete distribution used to
approximate standard normal. This function generate
order statistics of standard normal with same probability
assigned.

'approxNormalOrder2' also generate order statistics
of standard normal first. However, the probability
is calculated as normal kernal.
}
\details{
This function use \code{EnvStats::evNormOrdStats} to get
the order statisics of standard normal distribution. The
same probability is assigned for each order statistics.
}
\seealso{
\code{EnvStats::evNormOrdStats} for order
statisics of standard normal. \code{\link{fitMRMEapprox}}
for fit MRME with approximated measurement error.
}
\author{
Chaoran Hu
}
