% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movres.R
\name{estVarMRME_Godambe}
\alias{estVarMRME_Godambe}
\alias{estVarMRMEnaive_Godambe}
\title{Variance matrix of estimators from moving-resting process with measurement error}
\usage{
estVarMRME_Godambe(est_theta, data, nBS, integrControl = integr.control())

estVarMRMEnaive_Godambe(est_theta, data, nBS,
  integrControl = integr.control())
}
\arguments{
\item{est_theta}{estimators of MRME model}

\item{data}{data used to process estimation}

\item{nBS}{number of bootstrap.}

\item{integrControl}{a list of control parameters for the \code{integrate}
function: rel.tol, abs.tol, subdivision.}
}
\value{
variance-covariance matrix of estimators
}
\description{
'estVarMRME_Godambe' uses Godambe information matrix to obtain variance matrix
of estimators from 'fitMRME'.

'estVarMRMEnaive_Godambe' use Godambe information matrix to obtain variance matrix
of estimators from 'fitMRME_naive'.
}
\examples{
\donttest{
tgrid <- seq(0, 10*100, length=100)
set.seed(123)
dat <- rMRME(tgird, 1, 0.5, 1, 0.01, "m")
estVarMRME_Godambe(c(1, 0.5, 1, 0.01), dat, nBS = 10)
estVarMRMEnaive_Godambe(c(1, 0.5, 1, 0.01), dat, nBS = 10)
}

}
\author{
Chaoran Hu
}
