% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{f109}
\alias{f109}
\title{GPS data of f109}
\format{
A data frame with 3917 rows and 3 variables:
\describe{
    \item{t}{time when the GPS coordinates were collected (unit: year)}
    \item{dE}{UTM easting (unit: meter)}
    \item{dN}{UTM northing (unit: meter)}
}
}
\usage{
f109
}
\description{
A dataset of GPS coordinates of a mature female mountain
lion living in the Gros Ventre Mountain Range near
Jackson, Wyoming. The data were collected by a code-only
GPS wildlife tracking collar from 2019 to 2012.
}
\keyword{datasets}
