% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{f109raw}
\alias{f109raw}
\title{GPS data of f109 (raw format)}
\format{
A data frame with 3917 rows and 3 variables:
\describe{
    \item{t}{Date and time when the GPS coordinates were collected (unit: year)}
    \item{dE}{Standardized UTM easting (unit: meter)}
    \item{dN}{Standardized UTM northing (unit: meter)}
}
}
\usage{
f109raw
}
\description{
The original format of `f109` dataset.
}
\keyword{datasets}
