% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalAnalysis_mr_bmme.R
\name{fitBmme.seasonal}
\alias{fitBmme.seasonal}
\title{Fit a bmme for seasonal analysis}
\usage{
fitBmme.seasonal(data, start = NULL, method = "Nelder-Mead", ...)
}
\arguments{
\item{data}{The dataset should be fitted with the same format
as the output of \code{seasonFilter}.}

\item{start, method}{These two params are the same as params
in \code{fitBmme}.}

\item{...}{The params will be passed to \code{optim}.}
}
\value{
A list of fit result, which is the same as \code{optim}'s
output.
}
\description{
Fit Brownian motion with measurement error for seasonal
analysis. (The document of this function should be merged
to \code{fitBmme}.)
}
