% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalAnalysis_smam.R
\name{fitMovRes.seasonal}
\alias{fitMovRes.seasonal}
\title{Fit a moving-resting model for seasonal analysis}
\usage{
fitMovRes.seasonal(data, start = NULL, likelihood = c("full", "composite"),
  logtr = FALSE, method = "Nelder-Mead", optim.control = list(),
  integrControl = integr.control())
}
\arguments{
\item{data}{The dataset should be fitted with the same format
as the output of \code{seasonFilter}.}

\item{start, likelihood, logtr, method, optim.control, integrControl}{The
same as the param from \code{smam::fitMovRes}.}
}
\value{
a list of estimation result, which is the same as the
output from \code{smam::fitMovRes}.
}
\description{
Fit a moving-resting model with embedded Brownian motion with
seasonal data. (The document of this function should be merged
to \code{smam::fitMovRes}.)
}
