% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMovResHun.R
\name{fitMovResHun.composite.parallel}
\alias{fitMovResHun.composite.parallel}
\title{Fit Three States Model (composite nllk with parallel feature)}
\usage{
fitMovResHun.composite.parallel(data, start, lower, upper, groupSize,
  numThreads = RcppParallel::defaultNumThreads() * 3/4,
  integrControl = integr.control())
}
\arguments{
\item{data}{The dataset should be fitted.}

\item{start}{The initial value for optimization.}

\item{lower}{Lower bound for optimization.}

\item{upper}{Upper bound for optimization.}

\item{groupSize}{int size of each group for composite likelihood.}

\item{numThreads}{Allocate the number of threads.}

\item{integrControl}{Integration control vector includes rel.tol,
abs.tol, and subdivisions.}
}
\value{
A list of estimation result.
}
\description{
Fit Model using COBYLA optimation approaches for composite negetive
log likelihood.
}
\author{
Chaoran Hu
}
