% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalAnalysis_mrh.R
\name{fitMovResHun.seasonal.parallel}
\alias{fitMovResHun.seasonal.parallel}
\title{Fit three states model for seasonal analysis}
\usage{
fitMovResHun.seasonal.parallel(data, start, lower = c(0.001, 0.001, 0.001,
  0.001, 0.001), upper = c(10, 10, 10, 10, 0.999),
  numThreads = RcppParallel::defaultNumThreads() * 3/4,
  integrControl = integr.control())
}
\arguments{
\item{data}{The dataset should be fitted, which have the same
format as the output of \code{seasonFilter}.}

\item{start, lower, upper, numThreads, integrControl}{All parameters
here are the same as \code{\link{fitMovResHun.parallel}}.}
}
\value{
A list of estimation result.
}
\description{
Fit model using COBYLA optimization approach for the seasonal
analysis. The special structure of data is require here, that
will be explained in detail below. The data is actually the
subsets of whole dataset which come from some certain period
of each year.
}
\seealso{
\code{\link{seasonFilter}}
}
\author{
Chaoran Hu
}
