% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalAnalysis_mr_bmme.R,
%   R/seasonalAnalysis_mrh.R
\name{fitBmme.seasonal}
\alias{fitBmme.seasonal}
\alias{fitMovRes.seasonal}
\alias{fitMovResHun.seasonal}
\title{Fit MRH, MR, BMME Models for Seasonal Analysis (seasonal analysis toolbox)}
\usage{
fitBmme.seasonal(data, start = NULL, method = "Nelder-Mead", ...)

fitMovRes.seasonal(data, start = NULL, likelihood = c("full", "composite"),
  logtr = FALSE, method = "Nelder-Mead", optim.control = list(),
  integrControl = integr.control())

fitMovResHun.seasonal(data, start, lower = c(0.001, 0.001, 0.001, 0.001,
  0.001), upper = c(10, 10, 10, 10, 0.999), numThreads = NULL,
  integrControl = integr.control())
}
\arguments{
\item{data}{The dataset should be fitted, which have the same
format as the output of \code{seasonFilter}.}

\item{start}{The initial value for optimization.}

\item{method, ...}{are the parameters passed to \code{optim}.}

\item{likelihood, logtr, optim.control}{are the same as \code{\link{fitMovRes}}.}

\item{lower, upper, numThreads, integrControl}{are the same as \code{\link{fitMovResHun}}.}
}
\value{
A list of estimation result with the same format as
corresponding normal functions.
}
\description{
Fit a MRH, MR or BMME models, that is a wrapper of \code{\link{fitMovResHun}},
\code{\link{fitMovRes}} and \code{\link{fitBmme}} separately for
seasonal analysis. The special structure of data is required,
that is the same as the return from \code{\link{seasonFilter}}.
During seasonal analysis, the data is actually the subsets of whole
dataset which come from some certain period of each year.
}
\seealso{
\code{\link{fitMovResHun}} for normal fit MRH model,
\code{\link{fitMovRes}} for normal fit MR model,
\code{\link{fitBmme}} for normal fit BMME model,
\code{\link{transfData}} for transfering original GPS data to standard
format, \code{\link{seasonFilter}} for subsetting given date interval
from transfered data.
}
\author{
Chaoran Hu
}
