% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMovResHun.R
\name{fitMovResHun1}
\alias{fitMovResHun1}
\alias{fitMovResHun5}
\alias{fitMovResHun.parallel}
\alias{fitMovResHun6}
\title{Fit Three States Model (forward algrithm nllk)}
\usage{
fitMovResHun1(data, start, optim.control = list(),
  integrControl = integr.control())

fitMovResHun5(data, start, lower, upper, integrControl = integr.control())

fitMovResHun.parallel(data, start, lower, upper, grainSize,
  numThreads = RcppParallel::defaultNumThreads() * 3/4,
  integrControl = integr.control())

fitMovResHun6(data, start, lower, upper, integrControl = integr.control())
}
\arguments{
\item{data}{The dataset should be fitted.}

\item{start}{The initial value for optimization.}

\item{optim.control}{The control parameter for R function optim.}

\item{integrControl}{Integration control vector includes rel.tol,
abs.tol, and subdivisions.}

\item{lower}{Lower bound for optimization.}

\item{upper}{Upper bound for optimization.}

\item{grainSize}{Minimum chunk size for parallelization.}

\item{numThreads}{Allocate the number of threads.}
}
\value{
A list of estimation result.
}
\description{
Fit Model using different optimation approaches. The fit1 bases on
Nelder-Mead. The fit5 bases on COBYLA. The fit6 bases on BOBYQA.
}
\author{
Chaoran Hu
}
