% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMovResHan.R
\name{fitMRH}
\alias{fitMRH}
\title{Fit a Moving-Resting-Handling Model with Embedded Brownian Motion}
\usage{
fitMRH(data, start, segment = NULL,
  numThreads = RcppParallel::defaultNumThreads() * 3/4, lower = c(0.001,
  0.001, 0.001, 0.001, 0.001), upper = c(10, 10, 10, 10, 0.999),
  integrControl = integr.control())
}
\arguments{
\item{data}{a \code{data.frame} whose first column is the observation
time, and other columns are location coordinates. If seasonal analysis
is required, additional columns is required to indicate how to subset
the dataset.}

\item{start}{The initial value for optimization, in the order of rate
of moving, rate of resting, rate of handling, volatility and switching
probability.}

\item{segment}{name of column which indicates how to subset the dataset.
This column should be an integer vector with zero
for discarding and non-zero number for each segments.}

\item{numThreads}{int, the number of threads allocated for parallel
computation. The default setup is 3/4 available threads. If this parameter
is less or equal to 1, the serial computation will be processed.}

\item{lower, upper}{Lower and upper bound for optimization.}

\item{integrControl}{Integration control vector includes rel.tol,
abs.tol, and subdivisions.}
}
\value{
A list of estimation result, with estimator, log-likelihood and
convergence code from \code{nloptr}.
}
\description{
Fit a Moving-Resting-Handling Model with Embedded Brownian Motion with
animal movement data at discretely observation times by maximizing
a full likelihood.
}
\examples{
\donttest{
## slow work, may take several hours
set.seed(06269)
tgrid <- seq(0, 400, by = 8)
dat <- rMRH(tgrid, 4, 0.5, 0.1, 5, 0.8, 'm')
fitMRH(dat, c(4, 0.5, 0.1, 5, 0.8)) ## parallel process
fitMRH(dat, c(4, 0.5, 0.1, 5, 0.8), numThreads = -1) ## serial process
}

}
\references{
Pozdnyakov, V., Elbroch, L.M., Hu, C., Meyer, T., and Yan, J. (2018+)
On estimation for Brownian motion governed by telegraph process with
multiple off states. <arXiv:1806.00849>
}
\seealso{
\code{\link{rMRH}} for simulation.
}
\author{
Chaoran Hu
}
