% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movres.R
\name{fitMRME}
\alias{fitMRME}
\title{Fit a Moving-Resting Model with Measurement Error}
\usage{
fitMRME(data, start, segment = NULL, method = "Nelder-Mead",
  optim.control = list(), integrControl = integr.control())
}
\arguments{
\item{data}{a data.frame whose first column is the observation time, and other
columns are location coordinates. If \code{segment} is not \code{NULL},
additional column with the same name given by \code{segment} should be
included. This additional column is used to indicate which part of
observations shoule be used to fit model. The value of this column can
be any integer with 0 means discarding this observation and non-0 means
using this obversvation. Using different non-zero numbers indicate different
segments. (See vignette for more details.)}

\item{start}{starting value of the model, a vector of four components
in the order of rate for moving, rate for resting, volatility, and
s.d. of Guassian measurement error.}

\item{segment}{character variable, name of the column which indicates segments,
in the given \code{data.frame}. The default value, \code{NULL}, means using
whole dataset to fit the model.}

\item{method}{the method argument to feed \code{optim}.}

\item{optim.control}{a list of control to be passed to \code{optim}.}

\item{integrControl}{a list of control parameters for the \code{integrate}
function: rel.tol, abs.tol, subdivision.}
}
\value{
a list of the following components:
\item{estimate}{the esimated parameter vector}
\item{loglik}{maximized loglikelihood or composite loglikelihood
evaluated at the estimate}
\item{convergence}{convergence code from \code{optim}}
}
\description{
Fit a Moving-Resting Model with Measurement Error. The measurement
error is modeled by Guassian noise. Using \code{segment} to fit part
of observations to the model. A practical application of this feature
is seasonal analysis.
}
\examples{
\donttest{
tgrid <- seq(0, 10, length=500)
set.seed(123)
## make it irregularly spaced
tgrid <- sort(sample(tgrid, 30)) # change to 400 for a larger sample
dat <- rMR(tgrid, 1, 0.5, 1, "m")
dat$X1 <- dat$X1 + rnorm(nrow(dat), 0, 0.01)
dat$X2 <- dat$X2 + rnorm(nrow(dat), 0, 0.01)

## fit whole dataset to the MR model
fit <- fitMRME(dat, start=c(1, 0.5, 1, 0.01))
fit


## fit part of dataset to the MR model
batch <- c(rep(0, 5), rep(1, 7), rep(0, 4), rep(2, 10), rep(0, 4))
dat.segment <- cbind(dat, batch)
fit.segment <- fitMRME(dat.segment, start = c(1, 0.5, 1, 0.01), segment = "batch")
head(dat.segment)
fit.segment
}

}
\references{
Hu, C., Pozdnyakov, V., and Yan, J. Moving-resting model with measurement
error. In process.
}
