% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_smam.R
\name{fitSmam}
\alias{fitSmam}
\title{Fit a Statistical Model of Animal Movement}
\usage{
fitSmam(data, model = c("bmme", "MovRes", "MovResHan"), start = NULL,
  likelihood = c("full", "composite"), logtr = FALSE, lower = c(0.001,
  0.001, 0.001, 0.001, 0.001), upper = c(10, 10, 10, 10, 0.999),
  method = "Nelder-Mead", optim.control = list(), ...,
  integrControl = integr.control(), numThreads = NULL)
}
\arguments{
\item{data}{The dataset should be fitted, which is a \code{data.frame} whose
first column is the observation time, and other columns are location coordinates.
If it is a \code{List} with the same format as the output of \code{seasonFilter},
the seasonal analysis will be processed.}

\item{model}{specify the statistical model be used, which should be
within \code{c("bmme", "MovRes", "MovResHan")}.}

\item{start}{starting value of the corresponding model. For bmme model, it should
be a vector of two component, one for sigma (sd of BM) and the other for delta
(sd for measurement error). If unspecified (NULL), a moment estimator will be used
assuming equal sigma and delta. For moving-resting model, it should be a vector of
three components in the order of rate for moving, rate for resting, and volatility.
For moving-resting-handling model, it should be a vector in order of rate
of moving, rate of resting, rate of handling, volatility and switching
probability.}

\item{likelihood}{a character string specifying the likelihood type to
maximize in estimation. This can be "full" for full likelihood or
"composite' for composite likelihood.}

\item{logtr}{logical, if TRUE parameters are estimated on the log scale.}

\item{lower, upper}{Lower and upper bound for optimization.}

\item{method}{the method argument to feed \code{optim}.}

\item{optim.control, ...}{a list of control to be passed to \code{optim}.}

\item{integrControl}{a list of control parameters for the \code{integrate}
function: rel.tol, abs.tol, subdivision.}

\item{numThreads}{int, the number of threads allocated for parallel
computation (for moving-resting-handling model). The default setup
is 3/4 available threads. (If this parameter is set to 1, serial algorithm
will be processed.)}
}
\value{
A list of estimation result with elements corresponding to selected
model.
}
\description{
Fit a statistcal model of animal movement, which can be
Brownian motion with measurement error (\code{"bmme"}),
moving-resting process with embedded Brownian motion
(\code{"MovRes"}), or moving-resting-handling process
with embedded Brownian motion (\code{"MovResHan"}).
}
\details{
This function \code{fitSmam} is a wrapper funtion of
\code{fitBmme}, \code{fitMovRes}, \code{fitMovResHan},
\code{fitMovResHan.parallel}, \code{fitBmme.seasonal},
\code{fitMovRes.seasonal}, and \code{fitMovResHan.seasonal}.
The parameters should be chosen according to the corresponding
function.
}
\examples{
set.seed(06269)
tgrid <- seq(0, 400, by = 8)

## fit bmme model
dat.bmme <- rSmam(tgrid, model = "bmme", sigma = 0.9, delta = 0.1)
fitSmam(dat.bmme, model = "bmme")

## fit moving-resting model
dat.movres <- rSmam(tgrid, model = "MovRes",
                    lamM = 1, lamR = 0.1,
                    sigma = 1000, s0 = "m")
fitSmam(dat.movres, model = "MovRes", start = c(1, 0.1, 1000),
        likelihood = "full")

## fit moving-resting-handling model
\donttest{
## slow work, may take several hours
dat.movreshan <- rSmam(tgrid, model = "MovResHan",
             lamM = 4, lamR = 0.5, lamH = 0.1,
             sigma = 5, p = 0.8, s0 = "m")
fitSmam(dat.movreshan, model = "MovResHan", start = c(4, 0.5, 0.1, 5, 0.8))
}

}
\references{
Pozdnyakov V., Meyer, TH., Wang, Y., and Yan, J. (2013)
On modeling animal movements using Brownian motion with measurement
error. Ecology 95(2): p247--253. doi:doi:10.1890/13-0532.1.

Yan, J., Chen, Y., Lawrence-Apfel, K., Ortega, I. M., Pozdnyakov, V.,
Williams, S., and Meyer, T. (2014) A moving-resting process with an
embedded Brownian motion for animal movements.
Population Ecology. 56(2): 401--415.

Pozdnyakov, V., Elbroch, L., Labarga, A., Meyer, T., and Yan, J.
(2017) Discretely observed Brownian motion governed by telegraph
process: estimation. Methodology and Computing in Applied Probability.
doi:10.1007/s11009-017-9547-6.

Pozdnyakov, V., Elbroch, L.M., Hu, C., Meyer, T., and Yan, J. (2018+)
On estimation for Brownian motion governed by telegraph process with
multiple off states. (Under Review)
}
\seealso{
\code{\link{fitBmme}}, \code{\link{fitMovRes}},
\code{\link{fitMovResHan}}, \code{\link{fitMovResHan.parallel}},
\code{\link{fitBmme.seasonal}}, \code{\link{fitMovRes.seasonal}},
and \code{\link{fitMovResHan.seasonal}} for original functions.
}
