% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movres.R
\name{integr.control}
\alias{integr.control}
\title{Auxiliary for Controlling Numerical Integration}
\usage{
integr.control(rel.tol = .Machine$double.eps^0.25, abs.tol = rel.tol,
  subdivisions = 100L)
}
\arguments{
\item{rel.tol}{relative accuracy requested.}

\item{abs.tol}{absolute accuracy requested.}

\item{subdivisions}{the maximum number of subintervals.}
}
\value{
A list with components named as the arguments.
}
\description{
Auxiliary function for the numerical integration used in the
likelihood and composite likelihood functions. Typically only
used internally by 'fitMR' and 'fitMRH'.
}
\details{
The arguments are the same as \code{integrate}, but passed
down to the C API of Rdqags used by \code{integrate}.
}
