% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_smam.R
\name{rSmam}
\alias{rSmam}
\title{Sampling from a Statistical Model of Animal Movement}
\usage{
rSmam(time, model = c("bmme", "MovRes", "MovResHan"), lamM, lamR, lamH, sigma,
  delta, p, s0, dim = 2)
}
\arguments{
\item{time}{time points at which observations are to be simulated}

\item{model}{specify the statistical model be used, which should be
within \code{c("bmme", "MovRes", "MovResHan")}.}

\item{lamM}{rate parameter of the exponential duration while moving}

\item{lamR}{rate parameter of the exponential duration while resting}

\item{lamH}{rate parameter of the exponential duration while handling}

\item{sigma}{volatility parameter of the Brownian motion while moving}

\item{delta}{sd parameter of measurement error}

\item{p}{probability of choosing resting, and 1-p is probability of choosing handling}

\item{s0}{the state at time 0, must be one of "m" (moving) or "r" (resting/handling).}

\item{dim}{(integer) dimension of the Brownian motion}
}
\value{
A \code{data.frame} whose first column is the time points and whose
other columns are coordinates of the locations.
}
\description{
Sampling from a statistcal model of animal movement with
a given grid time. The specific model can be chosen from
Brownian motion with measurement error (\code{"bmme"}),
moving-resting process with embedded Brownian motion
(\code{"MovRes"}), and moving-resting-handling process
with embedded Brownian motion (\code{"MovResHan"}).
}
\details{
This function \code{rSmam} is a wrapper funtion of
\code{rbmme(time, dim, sigma, delta)},
\code{rMovRes(time, lamM, lamR, sigma, s0, dim)},
and \code{rMovResHan(time, lamM, lamR, lamH, sigma, p, s0, dim)}.
The parameters should be chosen according to the corresponding
model.
}
\examples{
tgrid <- seq(0, 10, length = 1001)
## make it irregularly spaced
tgrid <- sort(sample(tgrid, 800))

dat.bmme <- rSmam(tgrid, model = "bmme", sigma = 0.9, delta = 0.1)
## equivalent to rbmme(tgrid, sigma = 0.9, delta = 0.1)

dat.MovRes <- rSmam(tgrid, model = "MovRes",
                    lamM = 1, lamR = 0.1,
                    sigma = 1000, s0 = "m")
## equivalent to rMovRes(tgrid, lamM = 1, lamR = 0.1,
##                       sigma = 1000, s0 = "m")

dat.MovResHan <- rSmam(tgrid, model = "MovResHan",
                       lamM = 4, lamR = 0.4, lamH = 0.1,
                       sigma = 1000, p = 0.8, s0 = "m")
## equivalent to rMovResHan(tgrid, lamM = 4, lamR = 0.4, lamH = 0.1,
##                          sigma = 1000, p = 0.8, s0 = "m")


}
\references{
Pozdnyakov V., Meyer, TH., Wang, Y., and Yan, J. (2013)
On modeling animal movements using Brownian motion with measurement
error. Ecology 95(2): p247--253. doi:doi:10.1890/13-0532.1.

Yan, J., Chen, Y., Lawrence-Apfel, K., Ortega, I. M., Pozdnyakov, V.,
Williams, S., and Meyer, T. (2014) A moving-resting process with an
embedded Brownian motion for animal movements.
Population Ecology. 56(2): 401--415.

Pozdnyakov, V., Elbroch, L., Labarga, A., Meyer, T., and Yan, J.
(2017) Discretely observed Brownian motion governed by telegraph
process: estimation. Methodology and Computing in Applied Probability.
doi:10.1007/s11009-017-9547-6.

Pozdnyakov, V., Elbroch, L.M., Hu, C., Meyer, T., and Yan, J. (2018+)
On estimation for Brownian motion governed by telegraph process with
multiple off states. (Under Review)
}
\seealso{
\code{\link{rbmme}}, \code{\link{rMovRes}},
and \code{\link{rMovResHan}} for original functions.
}
