% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalAnalysis_mrh.R
\name{seasonFilter}
\alias{seasonFilter}
\title{Subsetting data during given season for each year (seasonal analysis toolbox)}
\usage{
seasonFilter(data, startDate, endDate)
}
\arguments{
\item{data}{The data be filtered, which has the same format
as the output from \code{\link{transfData}}.}

\item{startDate, endDate}{Start point and end point of
time interval during a year, which has the format "MM-DD".}
}
\value{
A \code{list} with each element is the one-year subset
of data, which is in the given time interval. For example,
suppose that we have data cross three years, this function
will return a list with three elements and each element
contains the subset of data of certain year.
}
\description{
Return subsets of data from each year, which is in given
time interval. The time interval here is for one year and
defined by \code{startDate} and \code{endDate}.
}
\author{
Chaoran Hu
}
