% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smam.R
\docType{package}
\name{smam}
\alias{smam}
\title{smam: Statistical Modeling of Animal Movements}
\description{
Animal movement models including moving-resting process
with embedded Brownian motion, Brownian motion with measurement error,
and moving-resting-handling process with embedded Brownian motion.
}
\author{
\itemize{
  \item *maintainer, author* Chaoran Hu \email{chaoran.hu@uconn.edu}
  \item *author* Vladimir Pozdnyakov \email{vladimir.pozdnyakov@uconn.edu}
  \item *author* Jun Yan \email{jun.yan@uconn.edu}
}
}
