% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thmam.R
\docType{package}
\name{thmam}
\alias{thmam}
\alias{thmam-package}
\title{thmam: Three States Model of Animal Movement}
\description{
Three states model of animal movement in R. This package is a
implementation of this model, and is used only for research work.
}
\author{
\itemize{
  \item *advisor* Jun Yan \email{jun.yan@uconn.edu}
  \item *advisor* Vladimir Pozdnyakov \email{vladimir.pozdnyakov@uconn.edu}
  \item *maintainer, author* Chaoran Hu \email{chaoran.hu@uconn.edu}
}
}
