% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{f109}
\alias{f109}
\title{GPS data of f109}
\format{
A data frame with 3919 rows and 4 variables:
\describe{
    \item{date}{Date when the GPS coordinates were collected}
    \item{cumTime}{Standardized time when the GPS coordinates were collected (unit: hour)}
    \item{centerE}{Standardized UTM easting (unit: km)}
    \item{centerN}{Standardized UTM northing (unit: km)}
}
}
\usage{
f109
}
\description{
A dataset of GPS coordinates of a mature female mountain
lion living in the Gros Ventre Mountain Range near
Jackson, Wyoming. The data were collected by a code-only
GPS wildlife tracking collar from 2009 to 2012.
}
\keyword{datasets}
