% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_mrh.R
\name{fitStateMRH}
\alias{fitStateMRH}
\title{Estimation of states at each time point with Moving-Resting-Handling Process}
\usage{
fitStateMRH(data, theta, integrControl = integr.control())
}
\arguments{
\item{data}{a \code{data.frame} whose first column is the observation
time, and other columns are location coordinates.}

\item{theta}{the parameters for Moving-Resting-Handling model, in the
order of rate of moving, rate of resting, rate of handling, volatility
and switching probability.}

\item{integrControl}{Integration control vector includes rel.tol,
abs.tol, and subdivisions.}
}
\value{
A \code{data.frame} contains estimated results, with elements:
\itemize{
 \item original data be estimated.
 \item conditional probability of moving, resting, handling (\code{p.m},
\code{p.r}, \code{p.h}), which is \eqn{Pr(S(t = t_k) = s_k | X)}.
 \item estimated states with 0-moving, 1-resting, 2-handling.
}
}
\description{
Estimate the state at each time point under the Moving-Resting-Handling
process with Embedded Brownian Motion with animal movement data at
discretely time points. The most likely state is estimated by maximizing
the probability of \eqn{Pr(S(t = t_k) = s_k | X)}, where X is the whole
data and \eqn{s_k} is the possible sates at \eqn{t_k} (moving, resting
or handling).
}
\examples{
\donttest{
set.seed(06269)
tgrid <- seq(0, 400, by = 8)
dat <- rMovResHan(tgrid, 4, 0.5, 0.1, 5, 0.8, 'm')
fitStateMRH(dat, c(4, 0.5, 0.1, 5, 0.8))
}

}
\references{
Pozdnyakov, V., Elbroch, L.M., Hu, C., Meyer, T., and Yan, J. (2018+)
On estimation for Brownian motion governed by telegraph process with
multiple off states. <arXiv:1806.00849>
}
\seealso{
\code{\link{rMovResHan}} for simulation.
\code{\link{fitMovResHan}} for estimation of parameters.
}
\author{
Chaoran Hu
}
