% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smam.R
\docType{package}
\name{smam}
\alias{smam}
\alias{smam-package}
\title{smam: Statistical Modeling of Animal Movements}
\description{
Animal movement models including
Moving-Resting Process with Embedded Brownian Motion,
Brownian Motion with Measurement Error,
Moving-Resting-Handling Process with Embedded Brownian Motion,
Moving-Resting Process with Measurement Error,
Moving-Moving Process with two Embedded Brownian Motions.
}
\author{
\itemize{
  \item *maintainer, author* Chaoran Hu \email{huchaoran.stat@gmail.com}
  \item *author* Vladimir Pozdnyakov \email{vladimir.pozdnyakov@uconn.edu}
  \item *author* Jun Yan \email{jun.yan@uconn.edu}
}
}
